# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Course Management System
  description: The main objective of the project is to facilitate the interaction between students and instructors related to the presentation of projects, tasks, thesis and allows instructors to give feedback to students.
servers:
  - url: 'http://192.168.1.8:5070/api'
    description: Home Server
  - url: 'http://localhost:5070/api'
    description: Development server
  - url: 'https://cms.kltpzyxm.live/api'
    description: Production server
paths:
...
  /user:
    get:
      summary: 'DEBUG: List all users in database'
      operationId: listUsers
      tags:
        - user
      security:
        - MasterKeyAuth: []
      responses:
        '200':
          description: 'success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfUsers'
        '400':
          description: 'bad request'   
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'  
    post:
      summary: 'add a new user to database.'
      operationId: addUser
      tags:
        - user
      responses:
        '201':
          description: 'user created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'bad request'   
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'  
        '403':
          description: 'user exists'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}:
    put:
      summary: 'replace user by id'
      operationId: replaceUser
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'user replaced'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusDetail'
        '400':
          description: 'bad request'   
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: 'update user by id'
      operationId: updateUser
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'user updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusDetail'
        '400':
          description: 'bad request'   
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      required:
        - _id
        - firstName
        - lastName
        - username
      properties:
        _id:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        nickname:
          type: string
        username:
          type: string
...
    ArrayOfUsers:
      type: array
      items:
        $ref: '#/components/schemas/User'
...
    StatusDetail:
      type: object
      properties:
        status:
          type: string
        detail:
          type: string
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth: 
      type: apiKey
      in: header
      name: X-API-KEY
    MasterKeyAuth: 
      type: apiKey
      in: header
      name: X-MASTER-KEY
security:
  # User Authentication
  - JWTAuth: []
  # Frontend/Backend exclusivity
  - ApiKeyAuth: []